pyproject.toml
[project]
name = "tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "ChugainovPavel"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "decouple (>=0.0.7,<0.0.8)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
